Refactor ndt_align_srv
