+++ localization/ndt_scan_matcher/CMakeLists.txt
@@ -33,7 +33,6 @@ ament_auto_add_executable(ndt_scan_matcher
   src/pose_array_interpolator.cpp
   src/tf2_listener_module.cpp
   src/map_module.cpp
-  src/pose_initialization_module.cpp
   src/map_update_module.cpp
 )
